Scroll display

+++ include/cursor.h
@@ -3,12 +3,13 @@
 
 #include <stdint.h>
 
-typedef uint32_t cursor_t;
+typedef int16_t cursor_t; // A type that describes a point on the screen
 
+// 2D screen cursor
 typedef struct Cursor_st
 {
-    cursor_t x;
-    cursor_t y;
+    cursor_t x; // X component. Bounds: [0, 480]
+    cursor_t y; // Y component. Bounds: [0, 272]
 } Cursor;
 
 #endif
